%p, hex implemented with bit shifting

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mshabano <mshabano@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 19:18:19 by mshabano          #+#    #+#             */
-/*   Updated: 2024/06/07 18:35:44 by mshabano         ###   ########.fr       */
+/*   Updated: 2024/06/07 21:33:02 by mshabano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ size_t	ft_strlen(const char *s);
 int put_str(const char *s, int len);
 int put_ptr(unsigned long ptr);
 int put_hex(long long n, char c);
-int put_int(int n);
+int put_int(long long n);
 int ft_printf(const char *s, ...);
 int print_f(const char **specifier, va_list args, int *printed);
 #endif
